Fixed timing issue especially in simulation

+++ ArduCopter/Copter.h
@@ -415,10 +415,10 @@ private:
     uint32_t brake_timeout_ms;
 
     // Delay the next navigation command
-    int32_t nav_delay_time_max;  // used for delaying the navigation commands (eg land,takeoff etc.)
-    uint32_t nav_delay_time_start = 0;
-    uint32_t nav_delay_abs_time_start = 0;
-    uint32_t nav_delay_debug_time_start = 0;
+    int64_t nav_delay_time_max;  // used for delaying the navigation commands (eg land,takeoff etc.)
+    uint64_t nav_delay_time_start = 0;
+    uint64_t nav_delay_abs_time_start = 0;
+    uint64_t nav_delay_debug_time_start = 0;
 
     // Flip
     Vector3f flip_orig_attitude;         // original copter attitude before flip
